refactor: split up test further
split up analysis specs for they need to be
kept separate when testing with promises
